+++ src/main/java/com/vaitilingom/projetbackend/services/formes/CubeService.java
@@ -1,7 +1,11 @@
 package com.vaitilingom.projetbackend.services.formes;
 
+import com.vaitilingom.projetbackend.models.auth.User;
+import com.vaitilingom.projetbackend.models.formes.Carre;
 import com.vaitilingom.projetbackend.repository.formes.CubeRepository;
 import com.vaitilingom.projetbackend.models.formes.Cube;
+import com.vaitilingom.projetbackend.services.auth.UserService;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,31 +15,44 @@ import java.util.Optional;
 public class CubeService {
     //Attribut
     private final CubeRepository cubeRepository;
+    private final UserService userService;
 
     //Constructeur
-    public CubeService(CubeRepository cubeRepository) {
+    public CubeService(CubeRepository cubeRepository, UserService userService) {
         this.cubeRepository = cubeRepository;
+        this.userService = userService;
     }
 
     //Méthodes CRUD
 
-    public List<Cube> getCubes() {
-        return cubeRepository.findAll();
+    public List<Cube> getCubesByUser(User user) {
+        return cubeRepository.findByUser(user);
     }
 
-    public Optional<Cube> getCubeById(int id) {
-        return cubeRepository.findById(id);
+    public Optional<Cube> getCubeById(int id, User user) {
+        Optional<Cube> cube = cubeRepository.findById(id);
+        if (cube.isPresent() && cube.get().getUser() != user) {
+            throw new AccessDeniedException("Vous n'avez pas le droit d'accéder à ce cube");
+        }
+        return cube;
     }
 
-    public Cube addCube(Cube cube) {
+
+    public Cube addCube(Cube cube, User user) {
+        cube.setUser(user);
         return cubeRepository.save(cube);
     }
 
-    public Cube updateCube(Cube cube) {
+    public Cube updateCube(Cube cube, User user) {
+        cube.setUser(user);
         return cubeRepository.save(cube);
     }
 
-    public void deleteCube(int id) {
+    public void deleteCube(int id, User user) {
+        Optional<Cube> cube = cubeRepository.findById(id);
+        if (cube.isPresent() && cube.get().getUser() != user) {
+            throw new AccessDeniedException("Vous n'avez pas le droit de supprimer ce cube");
+        }
         cubeRepository.deleteById(id);
     }
 
@@ -46,5 +63,11 @@ public class CubeService {
     }
     public double calculerVolume(Cube cube) {
         return cube.volume();
+    }
+
+    public Cube createCube(Cube cube, User user) {
+        cube.setUser(user);
+        Cube createdCube = cubeRepository.save(cube);
+        return createdCube;
     }
 }
