+++ src/main/java/com/vaitilingom/projetbackend/services/formes/SphereService.java
@@ -1,7 +1,11 @@
 package com.vaitilingom.projetbackend.services.formes;
 
+import com.vaitilingom.projetbackend.models.auth.User;
+import com.vaitilingom.projetbackend.models.formes.Cone;
 import com.vaitilingom.projetbackend.models.formes.Sphere;
 import com.vaitilingom.projetbackend.repository.formes.SphereRepository;
+import com.vaitilingom.projetbackend.services.auth.UserService;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,33 +15,45 @@ import java.util.Optional;
 public class SphereService {
     //Attribut
     private final SphereRepository sphereRepository;
+    private final UserService userService;
 
     //Constructeur
-    public SphereService(SphereRepository sphereRepository) {
+    public SphereService(SphereRepository sphereRepository, UserService userService) {
         this.sphereRepository = sphereRepository;
+        this.userService = userService;
     }
 
     //Méthodes CRUD
-    public List<Sphere> getSpheres() {
-        return sphereRepository.findAll();
+
+    public List<Sphere> getSpheresByUser(User user) {
+        return sphereRepository.findByUser(user);
     }
 
-    public Optional<Sphere> getSphereById(int id) {
-        return sphereRepository.findById(id);
+    public Optional<Sphere> getSphereById(int id, User user) {
+        Optional<Sphere> sphere = sphereRepository.findById(id);
+        if (sphere.isPresent() && sphere.get().getUser() != user) {
+            throw new AccessDeniedException("Vous n'avez pas le droit d'accéder à cette sphere");
+        }
+        return sphere;
     }
 
-    public Sphere addSphere(Sphere sphere) {
+    public Sphere addSphere(Sphere sphere, User user) {
+        sphere.setUser(user);
         return sphereRepository.save(sphere);
     }
 
-    public Sphere updateSphere(Sphere sphere) {
+    public Sphere updateSphere(Sphere sphere, User user) {
+        sphere.setUser(user);
         return sphereRepository.save(sphere);
     }
 
-    public void deleteSphere(int id) {
+    public void deleteSphere(int id, User user) {
+        Optional<Sphere> sphere = sphereRepository.findById(id);
+        if (sphere.isPresent() && sphere.get().getUser() != user) {
+            throw new AccessDeniedException("Vous n'avez pas le droit de supprimer cette sphere");
+        }
         sphereRepository.deleteById(id);
     }
-
     //Méthodes pragmatiques
 
     public double calculerSurface(Sphere sphere) {
@@ -46,5 +62,11 @@ public class SphereService {
 
     public double calculerVolume(Sphere sphere) {
         return sphere.volume();
+    }
+
+    public Sphere createSphere(Sphere sphere, User user) {
+        sphere.setUser(user);
+        Sphere createdSphere = sphereRepository.save(sphere);
+        return createdSphere;
     }
 }
